Added query string support
Updated the splitUrl function to check requests for query string then parse and create an object from that string. The application constructor will then pass it as a parameter to a constructor function in the Controller Class which populates a private $query_string variable.

Note: This includes my fix for the non-static openDatabaseconnection function.